+++ js/data/create-publication.js
@@ -1,0 +1,50 @@
+import {getRandomNumber} from '../utils.js';
+import {createComment} from './create-comment.js';
+
+const PHOTO_DESCRIPTIONS = [
+  'Будь таким человеком, с которым мечтаешь встретиться.',
+  'Будьте героями своих собственных историй!',
+  'Будьте счастливы в этот момент, потому что этот момент — и есть ваша жизнь.',
+  'Было бы все в этой жизни так же просто, как растолстеть.',
+  'В жизни я лучше, чем в инстаграме! 😉',
+  'В любой ситуации всегда улыбайтесь.',
+  'В море может и водится миллион рыб, но я единственная русалка.',
+  'В последнее время меня начинают преследовать умные мысли. Но я пока успеваю от них убежать…',
+  'В простоте есть удивительная красота.',
+  'В своей жизни ты должна играть только главную роль.',
+  'Вам нужно просто быть немножко сумасшедшим, ведь маленькое сумасшествие очень похоже на небеса.',
+  'Ваша скорость не имеет значения, пока вы продолжаете двигаться вперед.',
+  'На всякий случай, а то вдруг вы забыли, как я выгляжу…',
+  'Возможно, я не совершенна, но я определенно лимитированный экземпляр.',
+  'Сегодня у меня есть хороший шанс сказать вам: «Какой чудесный день!»',
+  'Время драгоценно, поэтому всегда тратьте его мудро.',
+  'Все мы рождаемся немного сумасшедшими, некоторые просто решают не меняться.',
+  'Бог должен быть невероятно креативным. Что я имею в виду? Просто посмотри на меня!',
+  'Все самое лучшее будет завтра! Но ты должна научиться ценить именно то, что имеешь уже сегодня.',
+  'Все только начинает становиться действительно хорошим.',
+  'Все что имеет значение — это просто быть собой.',
+  'Все, кем я хочу быть, это те, кем я была.',
+  'Все, что мне нужно, чтобы быть счастливой, — это шесть месяцев отпуска… Дважды в год.',
+  'Всегда будьте лучшим вариантом для себя.',
+  'Всегда держите голову высоко, не взирая на трудности и невзгоды.',
+];
+
+const LIKES_COUNT_MIN = 15;
+const LIKES_COUNT_MAX = 200;
+
+const COMMENTS_COUNT_MIN = 0;
+const COMMENTS_COUNT_MAX = 4;
+
+const createPublication = (i) => {
+  const commentsCount = getRandomNumber(COMMENTS_COUNT_MIN, COMMENTS_COUNT_MAX);
+
+  return {
+    id: i + 1,
+    url: `photos/${i + 1}.jpg`,
+    description: PHOTO_DESCRIPTIONS[i],
+    likes: getRandomNumber(LIKES_COUNT_MIN, LIKES_COUNT_MAX),
+    comments: Array.from({length: commentsCount}, createComment),
+  };
+};
+
+export {createPublication};
